+++ src/compliance_tests/TestIso_7_4_5.cpp
@@ -77,7 +77,7 @@ class TestIso_7_4_5 : public test::TestBase
         void ConfigureTest()
         {
             FillTestVariants(VariantMatchType::CommonAndFd);
-            for (int i = 0; i < 3; i++)
+            for (size_t i = 0; i < 3; i++)
             {
                 AddElemTest(TestVariant::Common, ElemTest(i + 1, FrameKind::Can20));
                 AddElemTest(TestVariant::CanFdEna, ElemTest(i + 1, FrameKind::CanFd));
@@ -110,7 +110,7 @@ class TestIso_7_4_5 : public test::TestBase
             mon_bit_frm->InsertOvrlFrm(0, BitKind::Interm);
             drv_bit_frm->InsertOvrlFrm(0, BitKind::Interm);
 
-            int bit_to_corrupt;
+            size_t bit_to_corrupt;
             if (elem_test.index_ == 1)
                 bit_to_corrupt = 2;
             else if (elem_test.index_ == 2)
@@ -121,7 +121,7 @@ class TestIso_7_4_5 : public test::TestBase
             TestMessage("Forcing Overload delimiter bit %d to recessive", bit_to_corrupt);
 
             Bit *bit = drv_bit_frm->GetBitOf(bit_to_corrupt - 1, BitKind::OvrlDelim);
-            int bit_index = drv_bit_frm->GetBitIndex(bit);
+            size_t bit_index = drv_bit_frm->GetBitIndex(bit);
             bit->val_ = BitVal::Dominant;
 
             drv_bit_frm->InsertActErrFrm(bit_index + 1);
